refactor: keep feature policy in table diff
