+++ src/delta_engine/domain/model/__init__.py
@@ -34,7 +34,7 @@ from delta_engine.domain.model.table import (
     TableAspect,
     TableKind,
 )
-from delta_engine.domain.model.table_feature import TableFeature, required_table_features
+from delta_engine.domain.model.table_feature import TableFeature
 
 __all__ = [
     "ALL_ASPECTS",
@@ -69,5 +69,4 @@ __all__ = [
     "TimestampNtz",
     "Variant",
     "key_signature",
-    "required_table_features",
 ]
